+++ JNI/jni_algorithm_biermann_minisat.h
@@ -10,10 +10,10 @@ extern "C" {
 /*
  * Class:     de_libalf_jni_AlgorithmBiermannMiniSAT
  * Method:    init
- * Signature: (II)I
+ * Signature: (JI)J
  */
-JNIEXPORT jint JNICALL Java_de_libalf_jni_AlgorithmBiermannMiniSAT_init
-  (JNIEnv *, jobject, jint, jint);
+JNIEXPORT jlong JNICALL Java_de_libalf_jni_AlgorithmBiermannMiniSAT_init
+  (JNIEnv *, jobject, jlong, jint);
 
 #ifdef __cplusplus
 }
